extracts omniauth testing stub into own file
