+++ spec/rails_helper.rb
@@ -6,6 +6,7 @@ require File.expand_path('../../config/environment', __FILE__)
 abort("The Rails environment is running in production mode!") if Rails.env.production?
 require 'rspec/rails'
 # Add additional requires below this line. Rails is not loaded until this point!
+require "support/omniauth.rb"
 require "support/factory_bot"
 require "capybara/rails"
 # Requires supporting ruby files with custom matchers and macros, etc, in
